refactor: Studio Service, OperatingHour 분리 (+CQRS에 따라 네이밍 수정)
